Report errors when loading YAML files with duplicate keys (#257)
* Report errors when loading YAML files with duplicate keys

This adds a custom constructor to the YAML loader created in
get_yaml_loader() so that duplicate keys in the YAML file result in a
PyYAML ConstructorError.

Duplicate keys are not allowed as per the YAML spec but are silently
ignored by PyYAML. This is a potential source of hard-to-detect errors
due to misconfigurations. It has been reported several times in the
PyYAML issues, e.g. https://github.com/yaml/pyyaml/issues/165. The
specific fix here was proposed at https://gist.github.com/pypt/94d747fe5180851196eb.